Fix filtering move filtering into main activity
