+++ app/src/main/java/pl/edu/amu/wmi/wmitimetable/service/ScheduleService.java
@@ -9,11 +9,9 @@ import java.util.Map;
 import java.util.TreeMap;
 import java.util.concurrent.TimeUnit;
 
-import pl.edu.amu.wmi.wmitimetable.model.Filter;
 import pl.edu.amu.wmi.wmitimetable.model.Meeting;
 import pl.edu.amu.wmi.wmitimetable.model.MeetingDay;
 import pl.edu.amu.wmi.wmitimetable.model.Schedule;
-import pl.edu.amu.wmi.wmitimetable.model.World;
 
 public class ScheduleService {
